% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_age.R
\name{calc_age}
\alias{calc_age}
\title{Title}
\usage{
calc_age(name, age)
}
\arguments{
\item{name}{(your name)}

\item{age}{(your age in years)}
}
\description{
Title
}
\examples{
calc_age(name = "Meagan", age = 23)
}
